requirements.txt
mysql-connector==2.2.9
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
pygame==2.5.2
pyzbar==0.1.9
